+++ KFXAbstracts/Classes/KFXCollectionViewController.h
@@ -9,8 +9,9 @@
  ************************************/
 
 
-
+// Cocoa Frameworks
 @import UIKit;
+@import MessageUI;
 // Categories
 #import <KFXAdditions/UIViewController+KFXAdditions.h>
 #import <KFXAdditions/UICollectionViewController+KFXAdditions.h>
@@ -18,13 +19,41 @@
 // Protocols
 #import <KFXAbstracts/KFXViewControllerNavigationDelegate.h>
 
-@interface KFXCollectionViewController : UICollectionViewController
+@interface KFXCollectionViewController : UICollectionViewController <MFMessageComposeViewControllerDelegate, MFMailComposeViewControllerDelegate>
 
 // Delegate
 @property (weak, nonatomic) id<KFXViewControllerNavigationDelegate> navigationDelegate;
 // Flags
 @property (nonatomic,getter=isVisible) BOOL visible;
 
+
+//--------------------------------------------------------
+#pragma mark Initilisation
+//--------------------------------------------------------
 +(instancetype)initiliseWithFlowLayout;
+
+
+//--------------------------------------------------------
+#pragma mark Item deselection
+//-------------------------------------------------------
+-(void)deselectAllItemsAnimated:(BOOL)animated;
+
+
+//--------------------------------------------------------
+#pragma mark Back Button
+//--------------------------------------------------------
+-(void)addCustomBackButtonWithImage:(UIImage*)image;
+
+//--------------------------------------------------------
+#pragma mark SMS & Email
+//--------------------------------------------------------
+-(void)presentEmailComposerWithSubject:(NSString*)subject
+                               message:(NSString*)message
+                         messageIsHTML:(BOOL)isHTML
+                           receipients:(NSArray<NSString*>*)recipients;
+-(void)presentSMSComposerWithSubject:(NSString*)subject
+                             message:(NSString*)message
+                         receipients:(NSArray<NSString*>*)recipients;
+
 
 @end