+++ KFXAbstracts/Classes/KFXCollectionViewController_Protected.h
@@ -1,0 +1,16 @@
+/********************************
+ *
+ * Copyright © 2016-2017 Christian Fox
+ * All Rights Reserved
+ * Full licence details can be found in the file 'LICENSE' or in the Pods-{yourProjectName}-acknowledgements.markdown
+ *
+ * This file is included with KFXAbstracts
+ *
+ ************************************/
+
+
+#import <KFXAbstracts/KFXAbstracts.h>
+
+@interface KFXCollectionViewController ()
+-(void)backButtonTapped:(id)sender;
+@end