+++ KFXAbstracts/Classes/KFXTableViewCell.h
@@ -21,7 +21,6 @@
 
 
 
-
 +(UINib*)nib;
 +(NSString*)reuseIdentifier;
 +(NSBundle*)bundle;