+++ KFXAbstracts/Classes/KFXTableViewController.h
@@ -11,7 +11,9 @@
 
 
 
+// Cocoa Frameworks
 @import UIKit;
+@import MessageUI;
 // Categories
 #import <KFXAdditions/UIViewController+KFXAdditions.h>
 #import <KFXAdditions/UITableViewController+KFXAdditions.h>
@@ -19,11 +21,36 @@
 // Protocols
 #import <KFXAbstracts/KFXViewControllerNavigationDelegate.h>
 
-@interface KFXTableViewController : UITableViewController 
+@interface KFXTableViewController : UITableViewController <MFMessageComposeViewControllerDelegate, MFMailComposeViewControllerDelegate>
 
 // Delegate
 @property (weak, nonatomic) id<KFXViewControllerNavigationDelegate> navigationDelegate;
 // Flags
 @property (nonatomic,getter=isVisible) BOOL visible;
+
+
+//--------------------------------------------------------
+#pragma mark Row deselection
+//--------------------------------------------------------
+-(void)deselectRowAnimated:(BOOL)animated;
+-(void)deselectAllRowsAnimated:(BOOL)animated;
+
+
+//--------------------------------------------------------
+#pragma mark Back Button
+//--------------------------------------------------------
+-(void)addCustomBackButtonWithImage:(UIImage*)image;
+
+//--------------------------------------------------------
+#pragma mark SMS & Email
+//--------------------------------------------------------
+-(void)presentEmailComposerWithSubject:(NSString*)subject
+                               message:(NSString*)message
+                         messageIsHTML:(BOOL)isHTML
+                           receipients:(NSArray<NSString*>*)recipients;
+-(void)presentSMSComposerWithSubject:(NSString*)subject
+                             message:(NSString*)message
+                         receipients:(NSArray<NSString*>*)recipients;
+
 
 @end