+++ KFXAbstracts/Classes/KFXViewController.h
@@ -10,7 +10,9 @@
 
 
 
+// Cocoa Frameworks
 @import UIKit;
+@import MessageUI;
 // Categories
 #import <KFXAdditions/UIViewController+KFXAdditions.h>
 #import <KFXAbstracts/UIViewController+KFXProgressHUD.h>
@@ -25,11 +27,37 @@
 //
 //@end
 
-@interface KFXViewController : UIViewController
+@interface KFXViewController : UIViewController <MFMessageComposeViewControllerDelegate, MFMailComposeViewControllerDelegate>
 
 // Delegate
 @property (weak, nonatomic) id<KFXViewControllerNavigationDelegate> navigationDelegate;
 // Flags
 @property (nonatomic,getter=isVisible) BOOL visible;
 
+
+//--------------------------------------------------------
+#pragma mark Back Button
+//--------------------------------------------------------
+-(void)addCustomBackButtonWithImage:(UIImage*)image;
+
+//--------------------------------------------------------
+#pragma mark SMS & Email
+//--------------------------------------------------------
+-(void)presentEmailComposerWithSubject:(NSString*)subject
+                               message:(NSString*)message
+                         messageIsHTML:(BOOL)isHTML
+                           receipients:(NSArray<NSString*>*)recipients;
+-(void)presentSMSComposerWithSubject:(NSString*)subject
+                             message:(NSString*)message
+                         receipients:(NSArray<NSString*>*)recipients;
+
+
 @end
+
+
+
+
+
+
+
+